+++ scripts/createBanner/makeBanner.mjs
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 export default ({name, repository, version}) =>
   `/*!
 =============================================================================