+++ src/getSubscriptionCount.ts
@@ -1,7 +1,7 @@
 import {EventStore} from './EventStore';
 import {NAMESPACE} from './constants';
 import {isEventized} from './isEventized';
-import { EventizedObject } from './types.js';
+import {EventizedObject} from './types.js';
 
 export const getSubscriptionCount = (o: object): number => {
   if (isEventized(o)) {